Batch processing rewrite, giving a 2x increase in speed.
Gets and calculates all the test values in a batch operation, according to the recommended best practise [1], so that expensive getCell and setValue calls are not made in the for-loop. Note that this will hide the test value outputs from the user as they are computed. The user can still see that the input and output cells in the sheet are changing though. But the user will have to wait a bit until all test values are printed out at the same time in the end. Speed increased by 2x in my test scenarios (based on 100 test values). This batch approach will also preserve the undo-history, so the user only has to press undo once to undo the entire table (and not once per value inserted into the table). The speed gain vs. hiding the intermittent outputting seems like a fair tradeoff that most people would make, and is especially important for larger data tables. Also, `SpreadsheetApp.getActive()` is deprecated. [1] https://developers.google.com/apps-script/guides/support/best-practices