Extensive help documentation for users and devs
Added extensive help documentation for users (so they will understand how to properly use WhatIfAnalysis.gs, and when not to use it in favor of a simpler and faster approach). Used ES6 template literal syntax to do that, but Google Sheets runs on Google V8 engine and successfully manages ES6. Also added some documentation for developers so that the two main classes of Data Tables are distinguished and made very clear. Number of columns could be confusing, so made that explicit.