Fix critical bug in 2D (bivariate) Data Table calculation (#12)
* Fix critical bug in 2D (bivariate) Data Table calculation

In the 2D matrix calculation, the setting of the `rowinput` and `colinput` cells were swapped in the for-loop, giving wrong output values set in the data table. Example test case here, with expected output (which can be recreated with the given test calculation formula `=ROWINPUT_CELL+(COLINPUT_CELL*2)`): https://docs.google.com/spreadsheets/d/1IE07D9XeIJou7LIh55L0UPkzp_7M-Kono_fdCtOo7dQ/edit?usp=sharing

* Batch processing rewrite, giving a 2x increase in speed.

Gets and calculates all the test values in a batch operation, according to the recommended best practise [1], so that expensive getCell and setValue calls are not made in the for-loop. Note that this will hide the test value outputs from the user as they are computed. The user can still see that the input and output cells in the sheet are changing though. But the user will have to wait a bit until all test values are printed out at the same time in the end. Speed increased by 2x in my test scenarios (based on 100 test values). This batch approach will also preserve the undo-history, so the user only has to press undo once to undo the entire table (and not once per value inserted into the table). The speed gain vs. hiding the intermittent outputting seems like a fair tradeoff that most people would make, and is especially important for larger data tables. Also, `SpreadsheetApp.getActive()` is deprecated. [1] https://developers.google.com/apps-script/guides/support/best-practices

* Extensive help documentation for users and devs

Added extensive help documentation for users (so they will understand how to properly use WhatIfAnalysis.gs, and when not to use it in favor of a simpler and faster approach). Used ES6 template literal syntax to do that, but Google Sheets runs on Google V8 engine and successfully manages ES6. Also added some documentation for developers so that the two main classes of Data Tables are distinguished and made very clear. Number of columns could be confusing, so made that explicit.

* Updated Readme.md with better examples

The previous examples were incomplete (didn't show the rowinput and colinput cells also needed). The bivariate example was misleading, as it had column headers with text instead of numeric values that the script actually requires.

* Flattened the test_values_array, which was a matrix but is expected to be an array. It previously worked regardless, since test_values were listed downwards in a column, but this should prevent future mishaps.

* Refactor code for legibility and reducing Range calls

Made the getRange calls more legible. Reduced some calls to Range methods because it presumably operates on the spreadsheet and that is expensive. Chaining several Range methods will presumably also amount to one interaction with the spreadsheet.

* Fix help text: B3 -> B2

Help text referenced B3 which was incorrect. B2 is correct.

* Added extensive help text

So that users will better know when to use WhatIfAnalysis and when to use other approaches.

* Add full help text to README.md

Since it is useful for people to know how to use, and when to use WhatIfAnalysis vs. other approaches. It represents general useful knowledge of efficiently dealing with complex data models in spreadsheets like Google Sheets. Having the help text only in the source code would hide it from anyone not peeking there, or having the script fully set up and working, and then coincidentally trigger the help text through e.g. clicking the help menu option.